+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # CHANGELOG for WordPress EDD Retroactive Licensing plugin
 
 ## master
-* Update description
+* Add email settings link
 * Enable EDD activation and version checking
+* Removed unused EDD_Retroactive_Licensing_Settings
+* Update description
+* Use `get_edd_options` and `set_edd_options` methods
 
 ## 0.0.1
 * Initial code release